Move sound and haptic to utils package
