+++ app/src/test/java/com/nqueens/game/features/nqueens/ui/state/NQueensBoardUiStateTest.kt
@@ -8,8 +8,8 @@ import com.nqueens.game.core.board.domain.games.GameState
 import com.nqueens.game.core.board.domain.pieces.PieceColor
 import com.nqueens.game.core.board.domain.pieces.QueenPiece
 import com.nqueens.game.core.board.ui.state.SelectedState
-import com.nqueens.game.core.haptic.DummyHapticFeedbackManager
-import com.nqueens.game.core.sound.DummySoundManager
+import com.nqueens.game.core.utils.haptic.DummyHapticFeedbackManager
+import com.nqueens.game.core.utils.sound.DummySoundManager
 import com.nqueens.game.features.nqueens.domain.NQueensBoardGame
 import kotlinx.coroutines.test.runTest
 import org.assertj.core.api.Assertions.assertThat
